events: the slab walk is the match engine — drop the cursor tree and the descending union
