+++ cmd/stellar-rpc/internal/rpcv2/stores/event/slab_match_bench_test.go
@@ -1,35 +1,36 @@
 package event
 
-// slab_match_bench_test.go measures the slab engine against the cursor tree at
-// two levels, because they answer different questions:
+// slab_match_bench_test.go measures the match path at three levels, because
+// they answer different questions:
 //
-//   - BenchmarkMatchPage is what a getEvents page costs end to end — candidate
-//     generation plus the fetch and post-filter both engines share. It is the
-//     number a request sees, and the shared half dilutes the engine difference
-//     exactly as production does.
-//   - BenchmarkMatchCandidates strips the shared half and times only the
-//     machinery being replaced: the cursor tree and the descending union
-//     against the slab stepper.
+//   - BenchmarkMatchPage is what a getEvents page costs end to end: candidate
+//     generation plus the fetch and post-filter around it. It is the number a
+//     request sees.
+//   - BenchmarkMatchCandidates strips the fetch and times candidate generation
+//     alone, drained in the batch sizes the streaming loop uses.
+//   - BenchmarkCandidateSlab times candidate generation over synthetic term
+//     geometries instead of a corpus, so a shape can be posed directly: two
+//     fat terms overlapping thinly, six fat terms, ten single-term filters.
 //
-// Both run over the shaped corpus from slab_match_differential_test.go, sized
-// to span several real 65536-id slabs.
+// The first two run over the shaped corpus from slab_match_test.go, sized to
+// span several real 65536-id slabs.
 
 import (
 	"context"
+	"errors"
 	"iter"
 	"slices"
+	"sync"
 	"testing"
 
+	"github.com/RoaringBitmap/roaring/v2"
 	"github.com/stretchr/testify/require"
 
 	protocol "github.com/stellar/go-stellar-sdk/protocols/rpc"
-)
+	"github.com/stellar/go-stellar-sdk/xdr"
 
-// matchEngine is the shape both engines share, so a benchmark can name one
-// and drive it without branching in the timed loop.
-type matchEngine = func(
-	context.Context, Reader, []Filter, IDRange, bool, int,
-) iter.Seq2[Match, error]
+	"github.com/stellar/stellar-rpc/cmd/stellar-rpc/internal/rpcv2/chunk"
+)
 
 // benchDir is one direction arm of every case.
 type benchDir struct {
@@ -38,17 +39,6 @@ type benchDir struct {
 }
 
 func benchDirs() []benchDir { return []benchDir{{"asc", false}, {"desc", true}} }
-
-// benchArms pairs each engine with the name its benchmark reports under.
-func benchArms() []struct {
-	name   string
-	engine matchEngine
-} {
-	return []struct {
-		name   string
-		engine matchEngine
-	}{{"cursor", Matches}, {"slab", slabMatches}}
-}
 
 // benchCorpusSize spans four and a half slabs, so a page served from the head
 // of the window leaves most of the corpus untouched and a full scan crosses
@@ -91,7 +81,7 @@ type benchCase struct {
 func benchCases(f *shapedFixture) []benchCase {
 	return []benchCase{
 		{"common", f.filterCommonPage(), 1000},
-		{"spill", f.filterThinOverlap(), 1000},
+		{"overlap", f.filterThinOverlap(), 1000},
 		{"fullscan", f.filterDenseOnly(), 0},
 	}
 }
@@ -102,16 +92,14 @@ func BenchmarkMatchPage(b *testing.B) {
 	r := diffPostingsReader{diffReader{f.corpus}}
 	for _, c := range benchCases(f) {
 		for _, dir := range benchDirs() {
-			for _, arm := range benchArms() {
-				b.Run(c.name+"/"+dir.name+"/"+arm.name, func(b *testing.B) {
-					benchPageRun(b, r, arm.engine, c, dir.desc)
-				})
-			}
+			b.Run(c.name+"/"+dir.name, func(b *testing.B) {
+				benchPageRun(b, r, c, dir.desc)
+			})
 		}
 	}
 }
 
-func benchPageRun(b *testing.B, r Reader, engine matchEngine, c benchCase, desc bool) {
+func benchPageRun(b *testing.B, r Reader, c benchCase, desc bool) {
 	b.Helper()
 	ctx := context.Background()
 	window := IDRange{0, benchCorpusSize}
@@ -119,7 +107,7 @@ func benchPageRun(b *testing.B, r Reader, engine matchEngine, c benchCase, desc 
 	var sink uint32
 	for b.Loop() {
 		n := 0
-		for m, err := range engine(ctx, r, c.filters, window, desc, c.limit) {
+		for m, err := range Matches(ctx, r, c.filters, window, desc, c.limit) {
 			if err != nil {
 				b.Fatal(err)
 			}
@@ -132,9 +120,8 @@ func benchPageRun(b *testing.B, r Reader, engine matchEngine, c benchCase, desc 
 	_ = sink
 }
 
-// BenchmarkMatchCandidates times candidate generation alone: the cursor tree
-// and the descending materialized union against the slab stepper, drained in
-// the same batch sizes the streaming loops use, with no fetch and no
+// BenchmarkMatchCandidates times candidate generation alone: the slab stepper
+// drained in the batch sizes the streaming loop uses, with no fetch and no
 // post-filter in the way.
 func BenchmarkMatchCandidates(b *testing.B) {
 	f := benchFixture(b)
@@ -143,67 +130,22 @@ func BenchmarkMatchCandidates(b *testing.B) {
 
 	for _, c := range benchCases(f) {
 		for _, dir := range benchDirs() {
-			for _, name := range []string{"cursor", "slab"} {
-				b.Run(c.name+"/"+dir.name+"/"+name, func(b *testing.B) {
-					ctx := context.Background()
-					b.ReportAllocs()
-					var sink int
-					for b.Loop() {
-						if name == "slab" {
-							sink = drainSlabCandidates(ctx, b, r, c, window, dir.desc)
-						} else {
-							sink = drainCursorCandidates(ctx, b, r, c, window, dir.desc)
-						}
-					}
-					_ = sink
-				})
-			}
+			b.Run(c.name+"/"+dir.name, func(b *testing.B) {
+				ctx := context.Background()
+				b.ReportAllocs()
+				var sink int
+				for b.Loop() {
+					sink = drainSlabCandidates(ctx, b, r, c, window, dir.desc)
+				}
+				_ = sink
+			})
 		}
 	}
 }
 
-// drainCursorCandidates mirrors streamCandidates and streamUnion without the
-// fetch: same batch sizing, same per-batch id collection, same early stop.
-func drainCursorCandidates(
-	ctx context.Context, b *testing.B, r Reader, c benchCase, window IDRange, desc bool,
-) int {
-	b.Helper()
-	plans, keys, matchAll := planIndexTerms(c.filters)
-	if matchAll {
-		b.Fatal("benchmark filters must reach the index")
-	}
-	if desc {
-		union, err := unionForFilters(ctx, r, plans, keys, window)
-		if err != nil {
-			b.Fatal(err)
-		}
-		it := union.ReverseIterator()
-		return drainBatches(c, func(ids []uint32, batch int) []uint32 {
-			for it.HasNext() && len(ids) < batch {
-				ids = append(ids, it.Next())
-			}
-			return ids
-		})
-	}
-	sources, err := lookupPostings(ctx, r, keys)
-	if err != nil {
-		b.Fatal(err)
-	}
-	cand := candidateIter(plans, sources, window)
-	return drainBatches(c, func(ids []uint32, batch int) []uint32 {
-		for len(ids) < batch {
-			id, ok := cand.peek()
-			if !ok {
-				return ids
-			}
-			ids = append(ids, id)
-			cand.next()
-		}
-		return ids
-	})
-}
-
-// drainSlabCandidates is the same drain over the slab stepper.
+// drainSlabCandidates is the streaming loop's batch cadence with the fetch
+// removed: fill up to the batch size, stop when a fill comes back empty or the
+// page is full.
 func drainSlabCandidates(
 	ctx context.Context, b *testing.B, r Reader, c benchCase, window IDRange, desc bool,
 ) int {
@@ -217,20 +159,12 @@ func drainSlabCandidates(
 		b.Fatal(err)
 	}
 	st := newSlabStepper(plans, sources, window, desc)
-	return drainBatches(c, func(ids []uint32, batch int) []uint32 {
-		return st.appendUpTo(ids, batch)
-	})
-}
 
-// drainBatches is the streaming loops' batch cadence with the fetch removed:
-// fill up to the batch size, stop when a fill comes back empty or the page is
-// full. Both arms share it so the harness cannot favor either.
-func drainBatches(c benchCase, fill func(ids []uint32, batch int) []uint32) int {
 	batch, rest := batchSizes(c.limit)
 	ids := make([]uint32, 0, batch)
 	total := 0
 	for {
-		ids = fill(ids[:0], batch)
+		ids = st.appendUpTo(ids[:0], batch)
 		batch = rest
 		if len(ids) == 0 {
 			return total
@@ -242,14 +176,428 @@ func drainBatches(c benchCase, fill func(ids []uint32, batch int) []uint32) int 
 	}
 }
 
-// ───────────── the in-tree shape matrix, with a slab arm ─────────────
+// ───────────── the synthetic index ─────────────
 
-// BenchmarkCandidateSlab is the third arm of match_iter_test.go's per-shape
-// A/B: the same plan, the same postings, the same page fingerprint, answered
-// by the slab stepper instead of the cursor tree (BenchmarkCandidateTree) or
-// the whole-window union (BenchmarkCandidateMaterialized). The shape matrix
-// already holds the fat/thin-overlap geometries the alignment budget was built
-// for, so this is the directly comparable number.
+// stubIndex is a Reader over an in-memory mirror and one shared payload, so a
+// benchmark measures the match layer rather than the storage tier. FetchEvents
+// reuses its buffer, keeping the per-batch fetch cost equal in both directions.
+type stubIndex struct {
+	mirror *ConcurrentBitmaps
+	count  uint32
+	raw    []byte
+	buf    []Payload
+}
+
+func (s *stubIndex) ChunkID() chunk.ID           { return chunk.ID(0) }
+func (s *stubIndex) EventCount() (uint32, error) { return s.count, nil }
+
+func (s *stubIndex) Offsets() (*LedgerOffsets, error) {
+	return nil, errors.New("stubIndex: Offsets is not part of the match path")
+}
+
+func (s *stubIndex) LookupKeys(_ context.Context, keys []TermKey) ([]*roaring.Bitmap, error) {
+	out := make([]*roaring.Bitmap, len(keys))
+	for i, k := range keys {
+		bm, err := s.mirror.Get(k)
+		if err != nil {
+			return nil, err
+		}
+		out[i] = bm
+	}
+	return out, nil
+}
+
+func (s *stubIndex) FetchEvents(_ context.Context, ids []uint32) ([]Payload, error) {
+	if err := validateSortedEventIDs(ids); err != nil {
+		return nil, err
+	}
+	s.buf = s.buf[:0]
+	for range ids {
+		s.buf = append(s.buf, Payload{ContractEventBytes: s.raw})
+	}
+	return s.buf, nil
+}
+
+func (s *stubIndex) FetchRange(_ context.Context, start, count uint32) iter.Seq2[Payload, error] {
+	return func(yield func(Payload, error) bool) {
+		if err := validateFetchRange(start, count, s.count, s.ChunkID()); err != nil {
+			yield(Payload{}, err)
+			return
+		}
+		for range count {
+			if !yield(Payload{ContractEventBytes: s.raw}, nil) {
+				return
+			}
+		}
+	}
+}
+
+func (s *stubIndex) All(ctx context.Context) iter.Seq2[Payload, error] {
+	return s.FetchRange(ctx, 0, s.count)
+}
+
+// hotLikeIndex carries the same optional no-materialize seam HotStore does, so
+// the benchmarks exercise the production fast path (sparse terms read in
+// place) rather than the bitmap fallback.
+type hotLikeIndex struct{ *stubIndex }
+
+func (h *hotLikeIndex) lookupPostings(_ context.Context, keys []TermKey) ([]postings, error) {
+	out := make([]postings, len(keys))
+	for i, k := range keys {
+		out[i] = h.mirror.lookupPostings(k)
+	}
+	return out, nil
+}
+
+var (
+	_ Reader        = (*stubIndex)(nil)
+	_ Reader        = (*hotLikeIndex)(nil)
+	_ postingReader = (*hotLikeIndex)(nil)
+)
+
+const (
+	// ~4M events: half a production chunk (~9M), enough that the intermediates
+	// are the multi-container bitmaps a real chunk builds.
+	benchEvents = 1 << 22
+	benchPage   = 1000 // getEvents' max page size
+)
+
+type benchIndex struct {
+	reader  *hotLikeIndex
+	filters []Filter
+	window  IDRange
+}
+
+// newBenchIndex builds the synthetic chunk once for both directions: three
+// dense terms (one near-total, like the event type; two selective) plus a
+// long-tail sparse term below the mirror's promotion threshold, so the sparse
+// read path is on the plan.
+var newBenchIndex = sync.OnceValue(func() *benchIndex {
+	var contractA xdr.ContractId
+	contractA[0] = 0xA1
+	topic := xdr.ScSymbol("bench-topic")
+	topicVal := xdr.ScVal{Type: xdr.ScValTypeScvSymbol, Sym: &topic}
+	topicRaw, err := topicVal.MarshalBinary()
+	if err != nil {
+		panic(err)
+	}
+	ev := xdr.ContractEvent{
+		ContractId: &contractA,
+		Type:       xdr.ContractEventTypeContract,
+		Body: xdr.ContractEventBody{
+			V:  0,
+			V0: &xdr.ContractEventV0{Topics: []xdr.ScVal{topicVal}, Data: topicVal},
+		},
+	}
+	raw, err := ev.MarshalBinary()
+	if err != nil {
+		panic(err)
+	}
+
+	// Dense terms go in through the frozen-Bitmaps constructor (roaring mode);
+	// the sparse one goes in through AddTo so it stays under the promotion
+	// threshold and is stored as a plain id list.
+	bms := NewBitmaps()
+	typeKey := EventTypeTermKey(xdr.ContractEventTypeContract)
+	contractKey := ComputeTermKey(contractA[:], FieldContractID)
+	topic1Key := ComputeTermKey(topicRaw, FieldTopic1)
+	everything := make([]uint32, 0, benchEvents)
+	contractIDs := make([]uint32, 0, benchEvents/3+1)
+	topic1IDs := make([]uint32, 0, benchEvents/7+1)
+	for id := range uint32(benchEvents) {
+		everything = append(everything, id)
+		if id%3 == 0 {
+			contractIDs = append(contractIDs, id)
+		}
+		if id%7 == 0 {
+			topic1IDs = append(topic1IDs, id)
+		}
+	}
+	bms.AddTo(typeKey, everything...)
+	bms.AddTo(contractKey, contractIDs...)
+	bms.AddTo(topic1Key, topic1IDs...)
+	mirror := NewConcurrentBitmapsFromBitmaps(bms)
+
+	topic0Key := ComputeTermKey(topicRaw, FieldTopic0)
+	sparse := make([]uint32, 0, promotionThreshold-1)
+	for i := range uint32(promotionThreshold - 1) {
+		sparse = append(sparse, i*(benchEvents/promotionThreshold))
+	}
+	mirror.AddTo(topic0Key, sparse...)
+
+	eventType := xdr.ContractEventTypeContract
+	var topics [protocol.MaxTopicCount][]byte
+	topics[0] = topicRaw
+	return &benchIndex{
+		reader: &hotLikeIndex{&stubIndex{
+			mirror: mirror, count: benchEvents, raw: raw,
+		}},
+		filters: []Filter{
+			// Two dense groups AND-ed: the intersect arm.
+			{ContractID: contractA[:], EventType: &eventType},
+			// One long-tail sparse group: the arm Get used to materialize a
+			// bitmap for on every request.
+			{Topics: topics},
+		},
+		// A sub-window, so both window edges are live.
+		window: IDRange{Start: benchEvents / 4, End: benchEvents * 3 / 4},
+	}
+})
+
+// benchMatches drives one page-sized request and stops, the shape a getEvents
+// page actually has.
+func benchMatches(b *testing.B, descending bool) {
+	b.Helper()
+	fx := newBenchIndex()
+	ctx := context.Background()
+	b.ReportAllocs()
+	b.ResetTimer()
+	for b.Loop() {
+		n := 0
+		for _, err := range Matches(ctx, fx.reader, fx.filters, fx.window, descending, benchPage) {
+			if err != nil {
+				b.Fatal(err)
+			}
+			n++
+			if n == benchPage {
+				break
+			}
+		}
+		if n != benchPage {
+			b.Fatalf("fixture sanity: want %d matches, got %d", benchPage, n)
+		}
+	}
+}
+
+func BenchmarkMatchesAscending(b *testing.B)  { benchMatches(b, false) }
+func BenchmarkMatchesDescending(b *testing.B) { benchMatches(b, true) }
+
+// ───────────── the synthetic shape matrix ─────────────
+
+// benchFat is one fat term's cardinality against benchEvents: ~7% of the
+// domain, the density at which roaring holds a term as bitmap containers.
+const benchFat = 300_000
+
+// benchRand is a deterministic xorshift. The shapes must be identical from run
+// to run, and a fixed stride would hand a walking AND a regularity real
+// postings do not have.
+type benchRand uint64
+
+func (r *benchRand) next() uint64 {
+	x := uint64(*r)
+	x ^= x << 13
+	x ^= x >> 7
+	x ^= x << 17
+	*r = benchRand(x)
+	return x
+}
+
+// scatter draws k ascending ids from one residue class, one per stride at a
+// jittered offset. Terms on disjoint classes interleave at single-id
+// granularity while sharing nothing, so a shape's overlap is exactly the class
+// its terms share.
+func scatter(rng *benchRand, domain, m, res uint32, k int) []uint32 {
+	if k == 0 {
+		return nil
+	}
+	class := (domain - res + m - 1) / m
+	stride := class / uint32(k)
+	if stride == 0 {
+		panic("scatter: residue class too small for k")
+	}
+	ids := make([]uint32, k)
+	for t := range k {
+		ids[t] = (uint32(t)*stride+uint32(rng.next()%uint64(stride)))*m + res
+	}
+	return ids
+}
+
+// fatGroup builds n terms of card ids each, drawing private ids from one
+// residue class per term plus one class every term holds, so the joint
+// intersection is exactly that shared class.
+func fatGroup(rng *benchRand, domain, mod, base uint32, n, card, shared int) [][]uint32 {
+	common := scatter(rng, domain, mod, base+uint32(n), shared)
+	out := make([][]uint32, n)
+	for i := range out {
+		ids := scatter(rng, domain, mod, base+uint32(i), card-shared)
+		ids = append(ids, common...)
+		slices.Sort(ids)
+		out[i] = ids
+	}
+	return out
+}
+
+// benchShape is one synthetic candidate-set problem: a term corpus in the
+// mirror, the plan resolved over it, and the page the engine must produce.
+type benchShape struct {
+	name   string
+	reader *hotLikeIndex
+	plans  []termPlan
+	keys   []TermKey
+	window IDRange
+	// wantCount and wantSum fingerprint the first page. The benchmark checks
+	// them every iteration, so a harness that stopped answering the query
+	// cannot post a fast number.
+	wantCount int
+	wantSum   uint64
+}
+
+// newBenchShape indexes terms as one term each, resolves the window to most of
+// the domain with both edges live, and fingerprints the first page off the
+// naive materialized algebra in referenceCandidates.
+func newBenchShape(name string, domain uint32, terms [][]uint32, plans []termPlan) *benchShape {
+	bms := NewBitmaps()
+	keys := make([]TermKey, len(terms))
+	for i, ids := range terms {
+		keys[i] = TermKey{0: byte(i + 1)}
+		bms.AddTo(keys[i], ids...)
+	}
+	s := &benchShape{
+		name: name,
+		reader: &hotLikeIndex{&stubIndex{
+			mirror: NewConcurrentBitmapsFromBitmaps(bms), count: domain,
+		}},
+		plans:  plans,
+		keys:   keys,
+		window: IDRange{Start: domain / 32, End: domain - domain/32},
+	}
+	sources, err := lookupPostings(context.Background(), s.reader, s.keys)
+	if err != nil {
+		panic(err)
+	}
+	for _, id := range referenceCandidates(s.plans, sources, s.window) {
+		if s.wantCount == benchPage {
+			break
+		}
+		s.wantCount++
+		s.wantSum += uint64(id)
+	}
+	return s
+}
+
+// singleFilterPlan is one filter AND-ing n one-term groups: the intersect
+// shapes' plan.
+func singleFilterPlan(n int) []termPlan {
+	plan := make(termPlan, n)
+	for i := range plan {
+		plan[i] = []int{i}
+	}
+	return []termPlan{plan}
+}
+
+// benchShapes is the shape matrix, each entry built on first use so a -bench
+// selecting one shape pays for one shape. domain is a parameter so the
+// correctness twin of the matrix can run the same geometries small.
+func benchShapes(domain uint32) []struct {
+	name  string
+	build func() *benchShape
+} {
+	scale := func(n int) int { return max(1, n*int(domain)/benchEvents) }
+	fat := scale(benchFat)
+	// ~3% of a fat term: the partial overlap that makes an AND converge slowly
+	// without making it empty.
+	partial := fat * 3 / 100
+	// Just over one page once the window clips it: the intersection too small
+	// to fill a page early, so the walk spans the window.
+	tiny := scale(1200)
+
+	shapes := []struct {
+		name  string
+		build func() *benchShape
+	}{
+		{"a_and2_fat_3pct", func() *benchShape {
+			rng := benchRand(1)
+			return newBenchShape("a", domain,
+				fatGroup(&rng, domain, 3, 0, 2, fat, partial), singleFilterPlan(2))
+		}},
+		{"b_and3_fat_3pct", func() *benchShape {
+			rng := benchRand(2)
+			return newBenchShape("b", domain,
+				fatGroup(&rng, domain, 4, 0, 3, fat, partial), singleFilterPlan(3))
+		}},
+		{"c_and2_skew", func() *benchShape {
+			rng := benchRand(3)
+			// The small term is a subset of the fat one, spread over it, so the
+			// AND is entirely decided by the rare side.
+			big := scatter(&rng, domain, 1, 0, fat)
+			small := make([]uint32, 0, scale(2000))
+			step := len(big) / cap(small)
+			for i := range cap(small) {
+				small = append(small, big[i*step])
+			}
+			return newBenchShape("c", domain,
+				[][]uint32{big, small}, singleFilterPlan(2))
+		}},
+		{"d_and6_fat_tiny", func() *benchShape {
+			rng := benchRand(4)
+			return newBenchShape("d", domain,
+				fatGroup(&rng, domain, 7, 0, 6, fat, tiny), singleFilterPlan(6))
+		}},
+		{"e_or10_single_term", func() *benchShape {
+			rng := benchRand(5)
+			terms := fatGroup(&rng, domain, 11, 0, 10, scale(30_000), 0)
+			plans := make([]termPlan, len(terms))
+			for i := range plans {
+				plans[i] = termPlan{{i}}
+			}
+			return newBenchShape("e", domain, terms, plans)
+		}},
+		{"f_and2_fat_tiny", func() *benchShape {
+			rng := benchRand(6)
+			return newBenchShape("f", domain,
+				fatGroup(&rng, domain, 3, 0, 2, fat, tiny), singleFilterPlan(2))
+		}},
+		{"h_and2_fat_overlapping", func() *benchShape {
+			rng := benchRand(8)
+			// The serving default: one selective term AND-ed with a near-total
+			// one, so a page comes out of the window's first fraction.
+			selective := scatter(&rng, domain, 3, 0, fat)
+			nearAll := make([]uint32, 0, domain)
+			for id := range domain {
+				if id%50 != 7 {
+					nearAll = append(nearAll, id)
+				}
+			}
+			return newBenchShape("h", domain,
+				[][]uint32{selective, nearAll}, singleFilterPlan(2))
+		}},
+		{"g_and3_x4_filters", func() *benchShape {
+			rng := benchRand(7)
+			// Several filters, each AND-ing a few fat terms. Each filter owns
+			// four residue classes, so the filters overlap only where the union
+			// has to dedup them.
+			terms := make([][]uint32, 0, 12)
+			plans := make([]termPlan, 0, 4)
+			for f := range uint32(4) {
+				group := fatGroup(&rng, domain, 16, f*4, 3, scale(75_000), scale(2250))
+				plan := make(termPlan, len(group))
+				for i := range group {
+					plan[i] = []int{len(terms) + i}
+				}
+				terms = append(terms, group...)
+				plans = append(plans, plan)
+			}
+			return newBenchShape("g", domain, terms, plans)
+		}},
+	}
+	return shapes
+}
+
+// benchShapeCache keeps one built corpus per shape name. Benchmarks run one at
+// a time, so a plain map suffices.
+var benchShapeCache = map[string]*benchShape{}
+
+func shapeFor(name string, build func() *benchShape) *benchShape {
+	s, ok := benchShapeCache[name]
+	if !ok {
+		s = build()
+		benchShapeCache[name] = s
+	}
+	return s
+}
+
+// BenchmarkCandidateSlab pulls one page of candidates per shape, with the
+// fetch and the post-filter out of frame.
 func BenchmarkCandidateSlab(b *testing.B) {
 	for _, sh := range benchShapes(benchEvents) {
 		b.Run(sh.name, func(b *testing.B) {
@@ -286,10 +634,58 @@ func benchSlabPage(b *testing.B, s *benchShape) {
 	}
 }
 
-// TestBenchShapesAgreeSlab is TestBenchShapesAgree's slab twin: every geometry
-// the microbench above measures must be one the slab stepper answers exactly,
-// so a shape can never post a number for a query it gets wrong.
-func TestBenchShapesAgreeSlab(t *testing.T) {
+// referenceCandidates is an independent, naive materialized implementation of
+// the algebra the slab stepper answers: OR each group whole, AND the groups,
+// OR across filters, then clip to the window. It builds every intermediate at
+// full chunk width, which is what the stepper exists not to do, so agreement
+// between the two is a real check rather than a restatement.
+func referenceCandidates(plans []termPlan, sources []postings, window IDRange) []uint32 {
+	materialize := func(p postings) *roaring.Bitmap {
+		if bm := p.bitmap(); bm != nil {
+			return bm
+		}
+		bm := roaring.New()
+		bm.AddMany(p.ids)
+		return bm
+	}
+	union := roaring.New()
+	for _, plan := range plans {
+		var acc *roaring.Bitmap
+		missed := false
+		for _, slots := range plan {
+			group := roaring.New()
+			present := false
+			for _, s := range slots {
+				if sources[s].present() {
+					present = true
+					group.Or(materialize(sources[s]))
+				}
+			}
+			if !present {
+				missed = true
+				break
+			}
+			if acc == nil {
+				acc = group
+			} else {
+				acc.And(group)
+			}
+		}
+		if missed {
+			continue
+		}
+		union.Or(acc)
+	}
+	windowBM := roaring.New()
+	windowBM.AddRange(uint64(window.Start), uint64(window.End))
+	union.And(windowBM)
+	return union.ToArray()
+}
+
+// TestBenchShapesAgree runs the whole shape matrix small: every geometry the
+// microbench measures must be one the stepper answers exactly, in both
+// directions, so a shape can never post a number for a query it gets wrong.
+func TestBenchShapesAgree(t *testing.T) {
 	const domain = 1 << 16
 	for _, sh := range benchShapes(domain) {
 		t.Run(sh.name, func(t *testing.T) {
@@ -298,30 +694,27 @@ func TestBenchShapesAgreeSlab(t *testing.T) {
 			require.NoError(t, err)
 			want := referenceCandidates(s.plans, sources, s.window)
 
-			st := newSlabStepper(s.plans, sources, s.window, false)
-			got := []uint32{}
-			for {
-				before := len(got)
-				got = st.appendUpTo(got, before+512)
-				if len(got) == before {
-					break
-				}
-			}
+			got := drainStepper(newSlabStepper(s.plans, sources, s.window, false))
 			require.Equal(t, want, got)
 			require.NotEmpty(t, got, "shape sanity: the plan must select something")
 
 			// The descending arm reads the same set backwards.
-			rev := newSlabStepper(s.plans, sources, s.window, true)
-			gotDesc := []uint32{}
-			for {
-				before := len(gotDesc)
-				gotDesc = rev.appendUpTo(gotDesc, before+512)
-				if len(gotDesc) == before {
-					break
-				}
-			}
+			gotDesc := drainStepper(newSlabStepper(s.plans, sources, s.window, true))
 			slices.Reverse(gotDesc)
 			require.Equal(t, want, gotDesc)
 		})
+	}
+}
+
+// drainStepper pulls a stepper dry in 512-id fills, never returning nil so an
+// empty result compares equal to a materialized bitmap's ToArray().
+func drainStepper(st *slabStepper) []uint32 {
+	out := []uint32{}
+	for {
+		before := len(out)
+		out = st.appendUpTo(out, before+512)
+		if len(out) == before {
+			return out
+		}
 	}
 }
